+++ src/main/scala/epic/parser/models/PositionalNeuralModel.scala
@@ -21,7 +21,7 @@ import epic.dense.TanhTransform
 import epic.dense.OutputTransform
 import epic.dense.AffineOutputTransform
 import epic.dense.OutputEmbeddingTransform
-import epic.corefdense.Word2Vec
+import epic.dense.Word2Vec
 import scala.collection.mutable.HashMap
 import epic.dense.Word2VecSurfaceFeaturizerIndexed
 import epic.dense.Word2VecDepFeaturizerIndexed
